requirements.txt
django==3.2
markdown
django-ckeditor